Add temporary header
